requirements.txt
pandas
numpy
requests
scikit-learn
matplotlib
seaborn
networkx
psycopg2-binary
flask